refactor category and smart list APIs

+++ app/Actions/Api/SmartListMealAction.php
@@ -5,21 +5,33 @@ declare(strict_types=1);
 namespace App\Actions\Api;
 
 use App\DTOs\Api\SmartListMealData;
+use App\DTOs\Api\SmartListMealResult;
 use App\Models\SmartList;
 
 class SmartListMealAction
 {
-    public function execute(SmartList $smartList, SmartListMealData $data): SmartList
+    public function execute(SmartList $smartList, SmartListMealData $data): SmartListMealResult
     {
-        $smartList->meals()->syncWithoutDetaching([$data->mealId]);
-        return  $smartList->load('meals');
+        $attached = $smartList->meals()->syncWithoutDetaching([$data->mealId]);
+
+        $isNew = !empty($attached['attached']);
+
+        return new SmartListMealResult(
+            smartList: $smartList->load('meals')->loadCount('meals'),
+            isNew: $isNew
+        );
     }
 
-    public function remove(SmartList $smartList, int $mealId): SmartList
+    public function remove(SmartList $smartList, int $mealId): ?SmartList
     {
+        $exists = $smartList->meals()->where('meal_id', $mealId)->exists();
+        
+        if (!$exists) {
+            return null;
+        }
+
         $smartList->meals()->detach($mealId);
 
-        return $smartList->load('meals');
+        return $smartList->load('meals')->loadCount('meals');
     }
-
 }
